pyproject.toml
[project]
name = "nonebot-plugin-track-anime"
authors = [{name = "lbsucceed", email = "woaikanpika123@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
description = "a tracking anime tool"
version = "0.1.13"




[tool.poetry]
name = "nonebot-plugin-track-anime"
version = "0.1.13"
description = "anime track plugin for NoneBot2"
authors = ["lbsucceed"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/lbsucceed/nonebot-plugin-track-anime"
documentation = "https://github.com/lbsucceed/nonebot-plugin-track-anime/README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
nonebot2 = ">=2.2.0"
nonebot-adapter-onebot = "^2.4.4"
beautifulsoup4 = "^4.11.2"
playwright = "^1.28.1"
httpx = "^0.23.0"
imageio = "^2.22.0"


[project.urls]
Home = "https://github.com/lbsucceed/nonebot-plugin-track-anime"
"Bug Tracker" = "https://github.com/lbsucceed/nonebot-plugin-track-anime/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"